Added "Well" to the legend title in the flourescence trace.
Modified the Q2 Analyser so blank wells in the data or metadata are stripped out from the data frames.

Confirmed the code is now working to detect differences in wells between Data and Metadata, and not render the Respiration tab until it is resolved. Displays an error message if this is the case.